feat!: drop AGP 4.x support
