+++ README.md
@@ -268,7 +268,7 @@ The plugin is compatible with any working combination of these ranges:
 | Component             | Oldest | Newest        |
 |-----------------------|--------|---------------|
 | JDK                   | 11     | 21            |
-| Gradle                | 6.6.1  | 8.9           |
-| Android Gradle Plugin | 4.0.0  | 8.8.0-alpha01 |
+| Gradle                | 7.0    | 8.9           |
+| Android Gradle Plugin | 7.0.0  | 8.8.0-alpha01 |
 
 NOTE: only the latest of any prerelease versions (`alpha`, `beta`, `rc`) is supported.
