--- gradle-compat-6-1/build.gradle.kts
@@ -1,9 +1,0 @@
-plugins {
-  id("wtf.emulator.java")
-}
-
-dependencies {
-  compileOnly(libs.gradle.api.v61)
-
-  api(projects.gradleCompatApi)
-}
