--- gradle-compat-6-6/build.gradle.kts
@@ -1,10 +1,0 @@
-plugins {
-  id("wtf.emulator.java")
-}
-
-dependencies {
-  compileOnly(libs.gradle.api.v66)
-  api(projects.gradleCompat65)
-
-  api(projects.gradleCompatApi)
-}
