--- gradle-compat-6-8/build.gradle.kts
@@ -1,10 +1,0 @@
-plugins {
-  id("wtf.emulator.java")
-}
-
-dependencies {
-  compileOnly(libs.gradle.api.v68)
-  api(projects.gradleCompat66)
-
-  api(projects.gradleCompatApi)
-}
