+++ gradle-plugin-api/build.gradle.kts
@@ -5,7 +5,7 @@ plugins {
 }
 
 dependencies {
-  compileOnly(libs.gradle.api.v61)
+  compileOnly(libs.gradle.api.v70)
 
   compileOnly(libs.agp)
   compileOnly(libs.autovalue.annotations)
