+++ gradle/libs.versions.toml
@@ -11,12 +11,9 @@ lint = { id = "com.android.lint", version.ref = "lint" }
 gradle-plugins-publish = "com.vanniktech:gradle-maven-publish-plugin:0.29.0"
 gradle-plugins-lint = { module = "com.android.lint:com.android.lint.gradle.plugin ", version.ref = "lint" }
 
-agp = "com.android.tools.build:gradle:4.0.0"
+agp = "com.android.tools.build:gradle:7.0.0"
 
-gradle-api-v61 = "dev.gradleplugins:gradle-api:6.1.1"
-gradle-api-v65 = "dev.gradleplugins:gradle-api:6.5"
-gradle-api-v66 = "dev.gradleplugins:gradle-api:6.6"
-gradle-api-v68 = "dev.gradleplugins:gradle-api:6.8"
+gradle-api-v70 = "dev.gradleplugins:gradle-api:7.0"
 gradle-api-v74 = "dev.gradleplugins:gradle-api:7.4"
 
 lint-gradle = "androidx.lint:lint-gradle:1.0.0-alpha02"
